Lab6
This is the first part of our Othelo game lab. In this lab, we initialize a board from user-chosen size (2x2-26x26), allow for a game "scenario" be uploaded by placing as many tiles as needed anywhere on the board. It then displays the available moves for each colour and allows the user to input 1 legal move, making the move.