Lab7Part1
This is an extension of lab 6. In this lab, our Othello game is now functional. There is no set-up scenario and you just dive into a game. This version also does not display legal moves. We also implemented a basic "AI" opponent to play against which will make the move which flips the most tiles. In the even of a draw, it picks the one with the lowest row, then column values.